completeness of save Skill, in case of a new skill.

+++ gui/SkillGUI.py
@@ -1,6 +1,6 @@
 from PySide6.QtWidgets import QGraphicsItem, QGraphicsScene, QComboBox, QWidget, QGridLayout, QFileDialog, QListView, \
-    QGraphicsRectItem, QMainWindow, QGraphicsView, QLabel, QApplication, QLineEdit, QPushButton, QHBoxLayout, \
-    QVBoxLayout, QMessageBox, QGraphicsPixmapItem, QScrollArea, QCompleter, QTabWidget, QSplitter, QTextBrowser, QMenu
+    QGraphicsRectItem, QMainWindow, QGraphicsView, QLabel, QApplication, QLineEdit, QPushButton, QRadioButton, QCheckBox, QHBoxLayout, \
+    QVBoxLayout, QMessageBox, QGraphicsPixmapItem, QScrollArea, QCompleter, QTabWidget, QSplitter, QTextBrowser, QDialogButtonBox, QMenu
 from PySide6.QtCore import QPointF, Qt, QEvent, QRectF
 from PySide6.QtGui import QPainterPath, QPen, QColor, QPixmap, QBrush, QPainter, QTransform, QStandardItemModel, QImage, \
     QAction
@@ -1361,6 +1361,31 @@ class SkillGUI(QMainWindow):
         self.mainWidget.setLayout(self.layout)
         self.setCentralWidget(self.mainWidget)
 
+
+        self.saveSkillMessageBox = QMessageBox()
+
+        # Set the title and text of the message box
+        self.saveSkillMessageBox.setWindowTitle("Save Skill Dialog")
+
+        # Create widgets to add to the layout
+        self.saveSkMBCheckboxLocal = QCheckBox("Save To Local")
+        self.saveSkMBCheckboxCloud = QCheckBox("Save To Cloud")
+        self.saveSkMBCheckboxLocal.setChecked(True)
+        self.saveSkMBCheckboxCloud.setChecked(False)
+
+        # Add layout to the message box
+        self.saveSkillMessageBox.setCheckBox(self.saveSkMBCheckboxLocal)
+        # self.saveSkillMessageBox.setCheckBox(self.saveSkMBCheckboxCloud)
+        self.saveSkillMessageBox.setStandardButtons(QMessageBox.Cancel | QMessageBox.Ok)
+        self.saveSkillMessageBox.setDefaultButton(QMessageBox.Ok)
+
+        gridLayout =self.saveSkillMessageBox.layout()
+        gidx = gridLayout.indexOf(self.saveSkMBCheckboxLocal)
+        cbrow, cbcol, cbrow_span,cbcol_span = gridLayout.getItemPosition(gidx)
+
+        gridLayout.addWidget(self.saveSkMBCheckboxCloud, cbrow, cbcol+1, cbrow_span,cbcol_span)
+        self.saveSkMBCheckboxCloud.setVisible(True)
+
         # app = QApplication.instance()
         # screen = app.primaryScreen()
         # #print('Screen: %s' % screen.name())
@@ -2271,28 +2296,55 @@ class SkillGUI(QMainWindow):
 
     def save_skill_file(self):
         # bring out the load file dialog
-        sk_prefix = "win_chrome_amz_home"
-        skname = self.skFCWidget.skfc_infobox.get_skill_info().skname
-        my_skill_dir_path = app_info.app_home_path + "/resource/skills/my/" + sk_prefix + "/" + skname + "/scripts/"
-        if not os.path.exists(my_skill_dir_path):
-            os.makedirs(my_skill_dir_path)
-            print("Folder created:", my_skill_dir_path)
+        ret = self.saveSkillMessageBox.exec_()
+        if ret == QMessageBox.Yes:
+            sk_prefix = "win_chrome_amz_home"
+            skname = self.skFCWidget.skfc_infobox.get_skill_info().skname
+            my_skill_dir_path = app_info.app_home_path + "/resource/skills/my/" + sk_prefix + "/" + skname + "/scripts/"
+            my_skill_img_dir = app_info.app_home_path + "/resource/skills/my/" + sk_prefix + "/" + skname + "/images/"
+            if not os.path.exists(my_skill_dir_path):
+                os.makedirs(my_skill_dir_path)
+                print("Folder created:", my_skill_dir_path)
 
-        skd_file_path = my_skill_dir_path + skname + ".skd"
+            skd_file_path = my_skill_dir_path + skname + ".skd"
 
-        skd_data = self.skFCWidget.encode_json(indent=4)
-        if skd_file_path:
-            with open(skd_file_path, 'w') as file:
-                file.write(skd_data)
-                print(f'save skd file to {skd_file_path}')
+            skd_data = self.skFCWidget.encode_json(indent=4)
+            if skd_file_path:
+                with open(skd_file_path, 'w') as file:
+                    file.write(skd_data)
+                    print(f'save skd file to {skd_file_path}')
 
-        worksettings = self.get_work_settings()
-        psk_words = self.skFCWidget.skfc_scene.gen_psk_words(worksettings)
-        psk_file_path = my_skill_dir_path + skname + ".psk"
-        if psk_file_path:
-            with open(psk_file_path, 'w') as file:
-                file.write(psk_words)
-                print(f'save psk file to {psk_file_path}')
+            worksettings = self.get_work_settings()
+            psk_words = self.skFCWidget.skfc_scene.gen_psk_words(worksettings)
+            psk_file_path = my_skill_dir_path + skname + ".psk"
+            if psk_file_path:
+                with open(psk_file_path, 'w') as file:
+                    file.write(psk_words)
+                    print(f'save psk file to {psk_file_path}')
+
+            if self.saveSkMBCheckboxCloud.isChecked():
+                # save to cloud here.
+                print("saving this skill to cloud ")
+                upload_file(self.session, skd_file_path, self.cog.id_token, "skill")
+                upload_file(self.session, psk_file_path, self.cog.id_token, "skill")
+                # upload_file(self.session, csk_file_path, self.cog.id_token, "skill")
+
+                # upload
+                anchor_files = [f for f in os.listdir(my_skill_img_dir) if os.path.isfile(f)]
+                for anchor_file in anchor_files:
+                    upload_file(self.session, anchor_file, self.cog.id_token, "skill")
+
+                #add to cloud DB
+                new_skill = WORKSKILL(self.parent, skname)
+                # populate ts_skill here with these parameters:
+                # platform,app,site,page,name,path,main,descriptio,runtime
+
+                result = send_add_skills_request_to_cloud(self.session, [new_skill], self.cog.id_token)
+
+                # add skillManagerWin
+                self.parent.skills.add(new_skill)
+                self.parent.addSkillRowsToSkillManager()
+
 
     def cancel_run(self):
         #will add later a sure? dialog
